requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-appconf==1.0.6
django-import-export==4.1.1
django-select2==8.2.1
pillow==10.4.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
